requirements.txt
discord.py
git+https://github.com/SkOODaT/pyipa.git